Move all code to /src/lib.rs instead of resolver.rs for simplicity.
