+++ Cargo.toml
@@ -6,18 +6,19 @@ edition = "2018"
 
 [dependencies]
 base64 = "0.10.1"
+chrono = "0.4"
+cueball = { git = "https://github.com/joyent/rust-cueball.git" }
 derive_more = "0.14.0"
+futures = "0.1"
+quickcheck = "0.8.0"
+resolv-conf = "0.6.2"
 sha1 = "0.6.0"
 slog = "2.4.1"
 state_machine_future = "0.2.0"
-futures = "0.1"
 trust-dns-resolver = "0.10.3"
 trust-dns = "^0.13"
-chrono = "0.4"
 tokio = "0.1.5"
 tokio-core = "0.1.17"
-resolv-conf = "0.6.2"
-cueball = { git = "https://github.com/joyent/rust-cueball.git" }
 
 [dev-dependencies]
 slog-term = "2.4.0"
